# [js/webgpu] Optimize matmulnbits with M > 1