### Description
This PR is mainly to optimize decoding first token in phi3 model. For the first token, the matmulnbits is a matrix * matrix, which is very slow when the prompt message is very long, like inputs tokens > 450.

Both intel GPUs and NV GPUs see good improvement with this PR.

My test data is like below:
Decoding first token with input 499 tokens:
It becomes 4.9s from 6.2s on NV RTX 2000.
It becomes 28.8s from 52.6s on Intel UHD 770.