@sushanthr 2024-12-13
> @guschmue @sushraja-msft FYI. I notice @sushraja-msft already did some very good optimizations on webgpu native ep. I will port this PR to webgpu ep to make some comparisons.

Thanks JiaJia, jfyi there is this pending change as well https://github.com/microsoft/onnxruntime/pull/23071 that improves on that previous change. Looking forward to comparing performance.